pyproject.toml
[project]
name = "google-ca-playbooks-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "gunicorn>=23.0.0",
    "aiohttp>=3.13.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
